FEA: Do not repeat logs if they have exactly the same message as previously emitted logs
Most notably for warnings such as "Mood {mood} is missing from a list of known moods. Not critical, but colouring won't work on the entry." as it just clogged down the entire log list.
