+++ Pipfile
@@ -5,7 +5,7 @@ name = "pypi"
 
 [packages]
 rich = "~=14.2"
-typer = "~=0.21.1"
+typer = "~=0.24"
 
 [dev-packages]
 coverage = "~=7.6"
